Fixed issue with getPickBlock for dimension ids
Using the pickblock on Blockholes would return a blockhole to the overworld. Now it pick blocks the correct dimension. However, when you CTRL-Pickblock and it returns NBT, that block leads to the correct dimension but if you pickblock or CTRL-pickblock the +NBT placed block returns a different, perhaps random, dimension for some reason...