Renaming classes to make sense, tuner overhaul.
I had to break any current existing blocks by changing class names for most features so that it made more logical sense and was neater. I also overhauled the tuner by removing the tooltip CONTROLS and MODE function of it, instead opting for a @SubscribeEvent RenderGameOverlayEvent so it shows the controls in the top left corner of the screen when the Tuner is equipped and you hold shift. Tooltip still shows the mode and description. @SubscribeEvent shows mode all the time.